updated appc/spec and camlistore/ godeps to fix broken godep restore
