+++ Godeps/Godeps.json
@@ -44,8 +44,8 @@
 		},
 		{
 			"ImportPath": "github.com/appc/spec/schema",
-			"Comment": "v0.7.4",
-			"Rev": "0875c991a496fe2b65ed296551c667b199a6bfb8"
+			"Comment": "v0.7.4-6-g5d54e27",
+			"Rev": "5d54e27f1764a0309eafe12c9df7bac03f241646"
 		},
 		{
 			"ImportPath": "github.com/aws/aws-sdk-go/aws",
@@ -121,8 +121,8 @@
 			"Rev": "3e36af6d3af0e56fa3da71099f864933dea3d9fb"
 		},
 		{
-			"ImportPath": "github.com/camlistore/camlistore/pkg/errorutil",
-			"Rev": "9868aa0f8d8a93ff0b30ff0de46cc351b6b88b30"
+			"ImportPath": "github.com/camlistore/go4/errorutil",
+			"Rev": "3b6feb5cf32f573fb124905de36a54c24ee14ac1"
 		},
 		{
 			"ImportPath": "github.com/codegangsta/negroni",
