+++ Godeps/_workspace/src/github.com/appc/spec/schema/version.go
@@ -22,7 +22,7 @@ const (
 	// version represents the canonical version of the appc spec and tooling.
 	// For now, the schema and tooling is coupled with the spec itself, so
 	// this must be kept in sync with the VERSION file in the root of the repo.
-	version string = "0.7.4"
+	version string = "0.7.4+git"
 )
 
 var (
