--- Godeps/_workspace/src/github.com/camlistore/camlistore/pkg/legal/legal.go
@@ -1,50 +1,0 @@
-/*
-Copyright 2014 The Camlistore Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-// Package legal provides project-wide storage for compiled-in licenses.
-package legal
-
-var licenses []string
-
-func init() {
-	RegisterLicense(`
-Copyright 2014 The Camlistore Authors
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-`)
-}
-
-// RegisterLicense stores the license text.
-// It doesn't check whether the text was already present.
-func RegisterLicense(text string) {
-	licenses = append(licenses, text)
-	return
-}
-
-// Licenses returns a slice of the licenses.
-func Licenses() []string {
-	return licenses
-}
