+++ Godeps/_workspace/src/github.com/camlistore/go4/errorutil/highlight.go
